requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
dash
plotly
flask
sqlalchemy
pytest
jupyter
openpyxl
pymysql
python-dotenv
pyyaml
torch
xgboost
shap
dash-core-components
dash-html-components
dash-table
numpy
scikit-learn
statsmodels
dash_bootstrap_components
jwt 
ipinfo
